Reduce usage of HAVE_X11 in KRun and make it usable on non-X11
The diversion for some methods by having a KStartupInfoId is removed.
As KStartupInfoId is not an X11 specific class we do not need the
ifdefs anymore.

In addition the code which uses X11 specific functionality is adjusted
to be only used if we are on platform X11. For other platforms the
same fallback is used as already present through the ifdef diversion.

REVIEW: 115721