Use Q_OS_UNIX instead of HAVE_X11 to determine the platform we are on
We cannot properly determine the windowing system platform on unix
like systems in kprotocolmanager as it's not linking gui. Thus we
don't know whether we are on X11 or Wayland and there is no proper
way to figure it out, because both DISPLAY and WAYLAND_DISPLAY could
be defined.

As a solution we just force the platform to be always X11 when we
are on unix like systems (modulo mac).

REVIEW: 115723