Add a simple manual test for listing directory contents
This test uses KIO::listDir to list the contents of the path that is
provided as a command line argument. The test then waits for the user
to press Enter before it terminates, which makes it easy to measure
with external tools how much memory the test uses.

This commit is the 1st part of https://git.reviewboard.kde.org/r/115739/