Add a benchmark for UDSEntry
The benchmark measures how long typical UDSEntry operations take: add
fields to an entry, read fields from an entry, store a UDSEntryList in a
QByteArray, and read it back from the QByteArray.

All measurements are done both for UDSEntries with 8 fields (this is a
rather typical use case because kio_file usually creates 8 fields), and
for entries with the maximum number of fields.

This commit is the 2nd part of https://git.reviewboard.kde.org/r/115739/