Make UDSEntry a Q_MOVABLE type
This has the effect that a QList<UDSEntry> a.k.a. UDSEntryList will
store the actual entries in a single allocated block of memory, and not
pointers to UDSEntries which are allocated individually on the heap
(this means that this change is binary incompatible). This reduces the
memory usage by 32 bytes per UDSEntry in a QList because each memory
allocation uses at least 32 bytes on a 64-bit system.

This commit is the 3rd part of https://git.reviewboard.kde.org/r/115739/

REVIEW: 115739