create the timer on the heap with a parent
this will auto-migrate the timer with the KRun object when it is moved
between threads, allowing it to work in such situations.

several Sprinter plugins use this pattern so as to use KRun from a
worker thread (KRun needs to be moved to the main app thread during this
as it may create wigets, e.g. on failure)