Fix build
This reverts commit 1bdb286f62454bb13cc299abd45a297a56b45cea.

See http://build.kde.org/job/kio_master_qt5/94/console

The first problem was that a ')' is missing. After adding it, the
compiler complains again because QUrl does not have a protocol() method.

I do not know what the correct solution is, so I think that it is better
to revert the commit now to prevent that more people run into this
problem.

Please figure out a better solution and verify that it compiles before
committing.

CCMAIL: adawit@kde.org