Import from kio-extras
This is the beginning of revision history for the kurifilter plugin in
this module. If you want to look at revision history older than this,
please refer to the techbase wiki for how to use Git history grafting.
At the time of writing, this wiki is located here:

http://community.kde.org/Frameworks/GitOldHistory

Note that you need to graft the history of the kio-extras repository in
addition to grafting the history of the kdelibs repository.

If you have already performed the grafting and you don't see any
history beyond this commit, try running "git log" with the "--follow"
argument.

Branched from the kio-extras repo at commit
493876bb6c69748fed42c52a15403995ca199a6b