Fix crashes in KUrlNavigator that are caused by nested event loops
KUrlNavigator opens menus with exec() in a few places, and accesses
member variables or pointers to children after that. This can cause
crashes if the object has been deleted inside the nested event loops.
This is fixed by using QPointers to detect if an object was deleted
already, and return immediately in that case.

This is a forward-port of commit
f1196e8e9a94993e4d748d283d869c4ad205ff02 in kdelibs.

CCBUG: 293863