KDirLister: do not kill+restart running update jobs.
On a slow mounted directory, it can lead to killing many jobs and never
letting it time to finish, during a long download.

Instead, we schedule another update after the end of the update job.

This made testNewItems fail because it was triggering two updates due to
the multiple signals from KDirWatch. So postpone directory dirty signals
from KDirWatch, just like we did already with file dirty signals.

Change-Id: Ib536d7ecac2abe418ab1022f40c21dbe227a49ae