Fix passing of DESKTOP_STARTUP_ID to child process in kioexec
The code was hard disabled due to usage of Q_WS_X11. The task is
to get the startupid passed to kioexec and forward it to the
process started by kioexec. As Qt's xcb plugin removes the
DESKTOP_STARTUP_ID environment variable once it is read, this needs
special handling to setup the enivornment for the process to be
started.

Unfortunately the fix requires Qt 5.4 as the startupId is only
provided by QX11Extras since 5.4.

REVIEW: 121230