Fix test invocation when there's no /usr/bin/gs
The test assumed that an executable called /usr/bin/gs is present on the
system which is running the test. That's a can of worms, IMHO, as that
dependency is not specified anywhere, and users are in general free to
massage their $PATH and what not. There's certainly no /usr/bin/gs on a
CI system I'm playing with right now.

At the same time, the test appears to require at least one well-known
executable name using full path. Let's stick with /bin/sh, that has
quite high chance of being around on any given system.

Change-Id: I50ade4f5bde3193f0607d9b3c74592117d810192