Use two QVectors, rather than a QHash, internally in UDSEntry
The number of items in a UDSEntry is always so small that the O(1)
lookup that QHash provides is not necessary. Therefore, the
QHash<uint, Field> is replaced by a QVector<uint>, and a
QVector<Field>. Both QVectors contain the uint keys and the Field
values in exactly the same order, such that the position at which a
value is stored can be determined by scanning the QVector that contains
the keys.

This change can decrease the memory consumption of UDSEntry in typical
use cases by 50% because QVector stores the data in a much more compact
form than QHash, which allocates individual hash nodes for each
key/value pair on the heap.

Change-Id: I9c45415cee48a415b1cf3d4fbd1e6e5fcd2a3bb4
REVIEW: 118452