Support for ".hidden" files
This adds support for *.hidden* files to KDE.

When listing a directory, the files/folders listed in the *.hidden* file
will be hidden, unless the user has chosen to show hidden files.

This patch was initially based on the patch provided by Mark in Bug #246260.

Unittest by David Faure in next commit.

FEATURE: 246260
Change-Id: I5e28de50e1fb9f3517ac22c9f2160f50bf11f798