Apply the edited URL to the navigator when clicking the "Check" button
If the navigator is in "editable" mode, pressing Enter applies the
edited URL to the view. This commit also enables this behavior for the
case that the user clicks the button to the right of the line edit in
order to switch back to breadcrumb mode. Before this commit, the edited
URL was lost.

BUG: 257538
Change-Id: I553afb1f6af033a026e8d8cf08f57f682ac723c2
FIXED-IN: 5.7.0
CHANGELOG: Apply the entered URL to the view when clicking the button
that switches the URL navigator back to breadcrumb mode