Add fields method and deprecate listFields().
Internally we store the fields in a QVector, but we were only returning a QList (which is toList() on the internal QVector).
The listFields method is now deprecated in favor of the new fields method.

CHANGELOG: Deprecate UDSEntry::listFields and add the UDSEntry::fields method which returns a QVector without costly conversion.
Change-Id: I3c9fc18311112e2e8df10161c8826c9ed19e0947