Make it possible to issue a storedPut job with QIODevice data
This way we can call put with data that is not strictly in memory. This can
save memory usage as well as unneeded byte array copies.
Additionally, adopts it internally on the KIO::AccessManager class, which
was already receiving a QIODevice and was just calling ::readAll on it to
get everything in memory and then issuing the query.

REVIEW: 123443

CHANGELOG: Asynchronous QIODevice-based data fetch for KIO::storedPut and
KIO::AccessManager::put.