Make it possible to treat sequential QIODevice asynchronously
So far, we used to just read whenever some data was required. This worked
on sequential devices because the data is already available. This is not
the case when we have a sequential device, such as a socket, where data
arrives when it arrives. This will also prove useful on non-sequential
devices as well when we want to keep reading in case new data appears.

This patch takes the AsyncDataEnabled setting on accordingly by:
* Only reading from the device when readyRead is available.
* Finishing the transfer whenever the device is closed.

It also improves accessmanager by adopting the new functionality on
sequential KIO::AccessManager::put calls.
On the testing side, it will make sure that both KIO::AccessManager and
QNetworkAccessManager produce the same results.

REVIEW: 123514