Provide deterministic sort order for KDirSortFilterProxyModelPrivate::compare
We already provide a deterministic sort order for KDirSortFilterProxyModelPrivate::compare when not using the QCollator, but when using it we don't. This piggy-backs on the existing code to fix that.

The open dialog doesn't switch like crazy now when hovering between two folders with the same name and different case.

REVIEW: 124795