remove useless xdg-open shebang
the template is copied and edited by libkonq/dolphin and that editing in
fact looses the shebang anyway.
additionally the shebang would only make sense iff the template was
installed +x and that attribute was preserved.

if we want to make apps executable by default (which IMO would make sense)
we'd best had something like
> krun.cpp static bool makeServiceExecutable
run as part of the template copy/modification