add logic to use icons for default xdg user dirs
this acts as an additional fallback to mimetype iconing and .directory
overrides. xdg user dirs are relocatable and can change depending on user
locale as well as user configuration and additionally are used in a range
of different desktop environments making this a very runtime sort of
decision.

BUG: 352498
REVIEW: 125158
CHANGELOG: Using suitable icons for default folders in the user home
FIXED-IN: 5.16