Add autocompletion for '.' input which offers all hidden files/folders*
Currently KUrlCompletion only offers autocompletion for hidden folders when you
input at least one additional character after the dot.
With this patch all hidden folders will be offered when only a dot is present.

This behaviour is test covered. But the fix requires Qt >= 5.6.0.
(5.6.0-alpha doesn't have the fix, so the version check is for 5.6.1)

Change-Id: I2ef4bfdfe3eb681b9e0e71f7183047c79cd968b0
REVIEW: 125988
BUG: 354981