Fix wrong button in KUrlNavigator issue caused by 9dbe36f
9dbe36f734b5b839b2a6a934fad29d639e954498 breaks some assumption on
startIndex. The argument index passed to buttonUrl is based number of
'/'in full url. Because of that, url like "sftp://a@b.com/a/b" is now
shown as "sftp:a@b.com" / "b" / "b" in dolphin url bar.

This patch changes all relevant code that calls buttonUrl() to use
url.path() instead of url.toDisplayString() to count the number of "/".

REVIEW: 126295