kdirmodel: Fix improper (non-virt) dtor usage.
Noted by Coverity (CID 1019869), and could result in a set of
partially-destructed objects. Which, in this case, would probably leak
memory in the event of multiple levels of filesystem hierarchy from the
root KDirModel, but wouldn't cause any worse problems than that.

It is unfortunate there is no better way than adding a virtual call but
there are too many places we delete these nodes to make the fake-virtual
equivalent worth it.

REVIEW:126506