kio_http: fix porting bugs which broke caching.
CacheLocation was used instead of GenericCacheLocation, so the appname was inserted
into it which would have broken sharing the cache between apps. But anyway no
caching happened at all because the subdir of the cache wasn't being created
(missing mkpath).

Repairing the use of the cache brought up an assert just after writing, because
the size of cache entries wasn't the expected size; this was due to mtime -> QDateTime
porting. Repaired by converting to "seconds since epoch", like kdelibs4 was doing.
Milliseconds don't seem to be part of the RFCs for HTTP dates anyway.

Finally, rename the cache dir to kio_http to avoid namespace pollution,
~/.cache/http is too generic.

REVIEW: 126857

Change-Id: Ie13c9923f02429a8b1893ba9f66f733673d5ddad