Change DBus name of kio_http_cache_cleaner so it doesn't exit if the kde4 one is running.
This DBus service name is only used for "unique app" behaviour.
It's not used for actual DBus communication: the communication from kio_http
comes over a local socket. Since that socket is in ~/.kde/socket-$HOSTNAME
in kde4 and XDG_RUNTIME_DIR in KF5, the KF5 app can't talk to the
kdelibs4 cache cleaner, it does need to be able to start the KF5 instance.
The cache itself is in a different dir too anyway.

REVIEW: 126858

Change-Id: I474cccf0e1a6b52ddef9fd5f90c4c40b9e4d7504