kio_http: fix waiting until the cache cleaner listens to the socket.
The old code was doing connectToServer+waitForConnected, but this
fails immediately if connectToServer sets an error like QLocalSocket::ConnectionRefusedError.
waitForConnected() is only meaningful if we are in ConnectingState.

We do have to try multiple times (no way to wait until precisely
the time when it start listening), but not using waitForConnected,
so I used msleep(100) in the loop instead.

I could have used DBus activation of course (to block until it's ready),
but I assume DBus is not wanted here since that's not what's used for
communication in the first place.

REVIEW: 128685

Change-Id: Ib9ee4699a44daa3016a72a24c6e816b940e1ed5c