help: fix garbage sent when serving static files
Switch from a QByteArray to a simplier stack-based char array, using
the actual number of bytes read when sending the data using data()
(reusing the memory from the array).

This removes the garbage sent when serving static files, i.e. those
stored in directories and not as part of a .cache.bz2 file.

REVIEW: 126936