Fix NTLMv2 stage 3 response creation
The Blob structure occupies 28 bytes. Unfortunately on 64-bit
architectures the size grows to 32 bytes due to alignment. This causes
code using sizeof() to calculate its size to produce wrong data. The
immediate effect is that the target info is not copied correctly. In
order to fix this introduce a constant defining the desired size of the
blob structure.

Additionally the size of the target info was not calculated correctly
when extracting it from the phase 2 response.

REVIEW: 126982