Try NTLMv2 authentication if the server denies NTLMv1
Some IIS servers seem to be configured to reject NTLMv1 authentication
by refusing to reply to a NTLM stage 1 if the NTLMv2 flag is not set. If
such a thing happens try to send another stage 1 message with the NTLMv2
flag set and if the server accepts this continue with NTLMv2.

This also fixes invese logic when determining if the authentication
needs a password (it needs it during stage 3 response and not stage 1).

As a bonus this includes a test case for verifying NTLMv2 authentication
and a fix for one of the existing test cases which contained wrong
expected data (the expected response was generated without use of
username and password due to the inverse logic bug above).

REVIEW: 126990