Fix KFileItem::refresh() for symlinks. The wrong size, filetype and permissions were being set.
As I suspected by reading the code; and proven by new unittest code.
This fix
- removes one stat() call in refresh, given that init() already calls stat().
- shares the logic between the initial init() and later refresh()
and does not call the UDSDEntry(STATBUF) ctor in init(), so that we don't overwrite
any other UDS_ entries set by the kioslave.

The idea of storing everything in UDSEntry and removing KFileItem members is good,
but the UDSEntry ctor is not needed for that, and creates regressions.

CCMAIL: markg85@gmail.com