kshorturifilter: fix regression with mailto: not prepended when no mailer is installed.
The removal of kmailservice5 from kio meant that "mailto" is no longer
a "known protocol" when no email client is installed. But since this code
is applying regexps, when they match we can just go ahead, no need to
test that they point to a known protocol. This makes the behaviour more
stable, whether or not an email client is available.