KRun: add runApplication method.
Summary:
The main purpose compared to runService is to use QProcess directly rather
than KLauncher. In e.g. Plasma launchers, using KLauncher can block the
GUI for a long time, for no purpose since we're not actually going to make DBus
calls to it. In the long run, we could move towards DBus activation for such cases,
and runApplication otherwise.

While making a new method, I also took the opportunity to turn the bool tempFiles
to a QFlags.

The purpose of this is to be able to clean up the plasma fix https://phabricator.kde.org/D1898
one day.

Test Plan: Added a testcase to the unittest.

Reviewers: hein, broulik, davidedmundson, svuorela

Reviewed By: davidedmundson, svuorela

Subscribers: #plasma, #frameworks

Differential Revision: https://phabricator.kde.org/D1902