KIO: add SlaveBase::openPasswordDialogV2 for better error checking.
Instead of bool return value, V2 provides an int (job error) return value.
This allows to differenciate between "user canceled" and "error communicating
with kpasswdserver". This is important because upon "user canceled" the job
just finishes silently, so the user thinks the operation succeeded when in fact
it got silently canceled, very dangerous.

The actual communication error dated from when kpasswdserver was in kded
rather than kiod (and on a computer with kded5 not installed), but still,
let's make this robust to avoid a possible data loss situation.

REVIEW: 128289