KPropertiesDialog: add Checksums tab
This adds a Checksums tab to KPropertiesDialog.
The new tab allows the user to easily check the integrity of a (local) file, as
well as the ability to copy the file's checksums (MD5, SHA1 or SHA256) to the
clipboard.

Checksums are computed in a secondary thread and are cached to avoid duplicate
computations. Cached checksums are invalidated as soon as the file is modified
on disk.

REVIEW: 128283