KNewFileMenu: fix creating symlinks with a relative target.
KNameAndUrlInputDialog::url() makes URLs absolute (which is useful in
general, for e.g. passing to KIO), so I added urlText() to retrieve
the text as is.

This removes all the code about showing an error on remote URLs; we just
create symlinks to the requested target, whether it looks like "../foo" or
"a:/b" which could be a relative path too (although an unusual one).

Bug misreported as a comment in bug 352927.

REVIEW: 128451