KPropertiesDialog: simplify button box usage, fix behavior on Esc
slotCancel() isn't called when pressing Esc or closing the window with
the window manager button.

Doing this the more standard way (reimplementing accept() and reject())
simplifies the code (since KPageDialog sets up the buttonbox this way already)
and will be useful for knewfilemenutest which calls accept() on all types of dialog.

REVIEW: 128494