KNewFileMenu: Fix suggestion of new filename when file already exist in desktop
When creating a new file, if a file with the default name already exist then a
new name is suggested by KIO::suggestName. Now this works fine until the scheme
of file's url is "file" but in case the scheme is "desktop" (like when the url
in dolphin is set to desktop:/ or "Folder View Settings > Location" is set to
desktop:/) then the check for file's existence couldn't happen because
isLocalFile() was false. So KIO::suggestName was not called and no new filename
is suggested in case a file with the same filename already exist.

REVIEW: 128499