KUrlRequester: fix error box when selecting a file and reopening the file dialog.
"<foo> is a file, but a folder was expected" was shown because of the
setDirectoryUrl() call added in 580dfd4992ca7d1d4 to fix bug 364719.

This is nowadays unnecessary, apparently due to commit 18e62e7 in plasma-integration,
selectUrl() does the right thing including changing dirs.
The new unittest shows that this works, and that the error above doesn't
happen anymore.

REVIEW: 128675