Fix creating symlink over "desktop" protocol.
After some changes made to KNewFileMenu (RR 128527 and commit 1adef4b), creating symlink with "Create New->Basic Link..." in plasma folder view(location set to "desktop:/*") started giving error message.
The reason was that CopyJob was not stating the destination directory of the symlink. This patch fixes this issue.

REVIEW: 128618