Fix KFileItem::overlays returning empty string values
KFileItem::overlays always returns a string list with an empty string as the list's first element.
For example,
for a symlink the list will be like this: {"", emblem-symbiolic-link}
and for a plain text file it'll be like : {""} (here it should be completely empty, size=0)
This patch attempts to fix the issue.