KFileItemActions: add addPluginActionsTo(QMenu *).
This was missing in KFileItemActions so apps had to implement
this by themselves (this particular implementation is inspired by Dolphin's,
while KonqPopupMenu only had support for .desktop files and not JSON metadata).

This API will also be useful in folderview (which I'm porting away
from KonqPopupMenu) and kfind (same strategy).

REVIEW: 129019