kntlm: Distinguish between NULL and empty domain
The default KNTLM behaviour when the client supplies no domain name is
to use the target name provided by the server. This however prevents
from explicitly using an empty domain name.

The change makes use of the fact that QString can be either NULL or
empty and retains the old behaviour if the domain name is NULL (for
ex. "username"), but when a valid empty string is provided (for
ex. "\username") the empty domain name is used instead of the target
name.

This fixes some rare authentication failures for oddly configured servers.