Properly finish DropJobs when triggered is not emitted
Since QMenu doesn't have a 'canceled' signal that is emitted when the
menu is "rejected" (e.g. user clicks outside the menu),
we need to manually track whether an action has been triggered.
This way we can ensure to always emit the 'result' signal and
avoid hangs in applications.

https://bugreports.qt.io/browse/QTBUG-56761 tracks the status of a
possible 'canceled' signal in QMenu, which would supersed this commit.

REVIEW: 129262
BUG: 363936
FIXED-IN: 5.28