Never stretch the last (=date) column in the file dialog.
Stretching the date column makes it appear to need the excessive
width when adding up non-name column widths in expandNameColumn().
The name column consequently became too narrow.
Now why could the date column expand if the names needed a lot of
space? It probably has something to do with expandNameColumn()
being called several times while the directory model is loading
and the exact order in which things are updated. Needs a more
detailed analysis if this change causes a regression or still
doesn't completely fix the problem.

BUG: 312747