KRun: deprecate runUrl() in favor of runUrl() with RunFlags
CVE-2017-5330 shows that `runExecutables = true` can be a dangerous
default for the runUrl() function. We cannot change the default value to
false (while BIC, it would be a change of behavior), so we deprecate the
current runUrl() function in favor of a new runUrl() with a RunFlags
argument replacing the `tempFile` and `runExecutables` arguments.

This new argument cannot take a default value, otherwise the two
runUrl() signatures would be ambiguous and existing code
would not compile.

REVIEW: 129844