Restore Cancel button shortcut
"Ok" and "Cancel" buttons are commonly used with alt+o and alt+c
shortcuts. But currently alt+c shortcut is assigned to the
Checksums tab, which is very confusing especially when Checksums plugin
is not used for all types of files, so sometimes alt+c works as expected
other times it doesn't.

So let's revert back to the commonly used behaviour.

Differential Revision: https://phabricator.kde.org/D5023