kioexec: delegate upload to a kded module
Introduce a kded module that kioexec can use to watch the cached
files for changes. This allows kioexec to be fully functional even with
applications that fork on startup, like libreoffice.

If the kded module is up and running, kioexec tells it (via dbus)
which file should be watched for changes and where to upload it when
it is actually changed. All the files watched by the module are deleted
when the module is destroyed.

As a bonus side effect, the dialog that asks if changes should be uploaded
is now displayed every time the user saves the file.

If dbus is not available or the kded module is otherwise disabled,
then kioexec falls back on the old behavior.

BUG: 252026
BUG: 370532
FIXED-IN: 5.34

Differential Revision: D5030