make kiod5 an "agent" on Mac
This makes kiod5 build as an unbundled application with the LSUIElement
property set, so that it runs without GUI presence but can still post
functional dialogs when needed. The Mac OS doesn't allow dialogs to be
"attached" to other applications; to achieve a comparable effect kiod5
is brought to the foreground when it posts a password dialog.

Differential Revision: https://phabricator.kde.org/D7237