[kiocore] Check that klauncher is running or not
Summary:
We are currently checking if the klauncher is running under different
user, and if so, we decide to fork the slave.

In addition to that also check that klauncher is actually running or
not. if klauncher is not installed due to kinit being higher tier
framework then the kio or due to the usage of kio slave on !plasma
environments where kinit is very less likely to be installed.

This also helps packagers to avoid the circular dependency between kio
and kinit. For example,

- https://phabricator.kde.org/T4517
- https://bugs.launchpad.net/ubuntu/+source/kio/+bug/1504526
- https://bugs.archlinux.org/task/49249

Test Plan: Tested on openbox session where kdeinit5 was not functional

Reviewers: #frameworks, sitter, dfaure

Reviewed By: sitter, dfaure

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D7462