Add mimetype filtering capabilities to KUrlCompletion
Summary:
This feature already exists in QFileDialog, but is so far missing
from KCompletion. This in turn makes KUrlRequester's mimetype filter
look broken, as it "only" works for the dialog but not the lineedit!

Setting a mimetype filter list will exclude any files with a mime-
type name that is not contained in the filter list. Note that
directories are never matched against the mimetype filter.

With this patch applied, the filter is also properly applied via
KCompletion to the lineedit of a KUrlRequester.

Reviewers: dfaure

Subscribers: #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D7527