Add support for sharing file descriptor between file ioslave and it's KAuth helper
Sharing of file descriptor takes place when copying a file with root privileges.
First the helper will open the file with root privileges and then it will  send the
file descriptor to file ioslave. After that the slave will use the file descriptor to
read file's data or write data to it.

Differential Revision: https://phabricator.kde.org/D6709