Add ability to use the new kauth helper in file ioslave
This patch introduces various additions that will enable file ioslave to communicate
effectively with the kauth helper.
The most important of these additions is execWithElevatedPrivilege() method. Control is passed to this
method whenever there is an error and if the error is either EACCES or EPERM then it will call the helper
to perform the desired task as root.

Differential Revision: https://phabricator.kde.org/D6829