KIO::PreviewJob::defaultPlugins() function
Summary:
Adds a new KIO::PreviewJob::defaultPlugins() function that can lets clients get a list of thumbnail plugins that should be enabled by default (disabled plugins can still be enabled via the GUI in Dolphin).

The function returns a QStringList of all available plugins minus ones specified in an internal blacklist, because some plugins may not be useful enough to enable by default, but also are not useless enough to remove. Currently the only blacklisted plugin is the .txt file previewer, because at most (but not all) icon sizes, the preview is unreadably small and not very useful.

This supports D7440 and D8352.

Test Plan: Tested in KDE Neon. Function is available and works as expected.

Reviewers: #frameworks, broulik, #dolphin, markg

Reviewed By: markg

Subscribers: markg, anthonyfieroni, elvisangelaccio, #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D8347