Fix display of remote:/ in the qfiledialog
Summary:
remote: is designed to redirect every entry so it has an entry that
initially is defined as "remote:/network" then it changes to "network:/"
through setUrl which currently is passing QUrl().fileName() which is empty
for network:, smb: or mtp:.
This patch changes it so at least the displayed text isn't overriden if it
was coming from the UDS_DISPLAY_NAME entry in the UDS tuple we received.

Test Plan:
This behavior can be tested and confirmed using the qfiledialogtest in
plasma-integration/tests.

Reviewers: #frameworks, davidedmundson

Reviewed By: davidedmundson

Subscribers: ngraham

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D8643