Don't autoremove empty temp file
Summary:
Since 5.10 (commit 0dee566e98 in qtbase) Qt complains if we make a
system call with an empty filename (with a "Empty filename passed to
function" runtime warning).

I even hit an assert with my self-compiled build:

    ASSERT: "unnamedFile == d_func()->fileEntry.isEmpty()"

(in qtemporaryfile.cpp, line 514).

The problem is that after `socketfile.remove()` we get an empty
`fileName()` which will be a problem when the QTemporaryFile gets
destroyed at the end of the if() block.
To avoid this issue we disable the autoremove mode.

Test Plan:
No longer get the runtime warning every time a file item
gets hovered in KFileWidget.

Reviewers: dfaure

Subscribers: #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D8696