Hidding place groups implementation in KFilePlacesModel
Summary:
This give an implementation to the places model to hide an entire group
of places. The model is now holding group states as metadata properties
within the bookmarkmanager root object.

Depends on D8243
Depends on D8332
Depends on D8348
Depends on D8366

Depends on D8948
Depends on D8947
Depends on D8946
Depends on D8945
Depends on D8944
Depends on D8943
BUG: 300247

Test Plan:
Unit tested

{F5459711} (before remove "recently saved")

{F5459710} (after)

Reviewers: renatoo, ngraham, ervin, mwolff, mlaurent

Reviewed By: renatoo, ervin, mwolff, mlaurent

Subscribers: mwolff, ngraham, mlaurent, #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D8367