User can now hide an entire places group from KFilePlacesView
Summary:
Through the ui the user can now right click on a place group
and request to hide it.

Note: It is not the collapsing solution, which would need a way bigger
effort on the model to be achieve but an intermediary solution that could
be useful for users.

Depends on D8367

Depends on D8948
Depends on D8947
Depends on D8946
Depends on D8945
Depends on D8944
Depends on D8943

BUG: 300247

Test Plan:
Manual testing with the ui.

{F5459735} (before hidden section)

{F5459734} (menu for hiding section)

{F5459733} (after hidden section)

Reviewers: ngraham, renatoo, ervin, mwolff, mlaurent

Reviewed By: renatoo, ervin, mwolff, mlaurent

Subscribers: mwolff, #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D8450