Fix unintentional breadcrumb menu item activation
Summary:
Since breadcrumb menu is opened on mouse press, it may receive
the corresponding mouse release event which may unintentionally
activate the random item which happens to be under mouse pointer.

This commit makes the menu ignore the first mouse release event
unless the pointer was moved. It fixes the issue and still allows
'Press mouse button' -> 'Move to an item' -> 'Release the button'
usage scenario.

BUG: 380287
FIXED-IN: 5.44

Reviewers: #frameworks, dfaure

Reviewed By: dfaure

Subscribers: broulik, ngraham

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D8958